Split out command creation tools from utils.py into cmd.py
